Fix sidebar resizing for real
Sometimes the sidebar isn't resized after we apply our styles. This
happens when our script is injected after the entire cstimer app
has already loaded. We can force this condition to reproduce the
problem by changing our entrypoint to

    setTimeout(function(){
      (new TimerExtension(localStorage, fixedSettings)).setup();
    }, 1000)

The solution is to check the computed styles of the sidebar every
100ms until we can observe one of our style changes. Once we
detect our change we trigger a window resize event which cstimer
listens to so it can resize various UI elements.